Update README.md
Humoropedia.com is a media sharing site that aims to bring its users the best of funny jokes and inspirational quotes. It has a huge collection of famous quotes about life by many different authors. We are especially proud of our collection of Rumi quotes about love and life. Each of our quote collections has been assembled by hand. We aim to be different from all other sites that provide quotes and jokes because we aim to provide users with a much better user experience that works both on mobile and desktop. Anyone interested in working with us is welcome to contact us.